Simplify directory structure to have one dir
