--- src/scripts/cassandra_email/__init__.py
@@ -1,1 +1,0 @@
-from . import create_send_email, utils
