--- src/scripts/hunter_csv/__init__.py
@@ -1,1 +1,0 @@
-from . import create_hunter_csv
