--- src/scripts/hunter_csv/constants.py
@@ -1,4 +1,0 @@
-"""
-This Python file contains all constants used to generate a csv file for hunter to analyse to detect performance
-regressions.
-"""
